got navbar working
