+++ FrontEnd/myapp/package.json
@@ -5,6 +5,8 @@
   "dependencies": {
     "@emotion/react": "^11.8.2",
     "@emotion/styled": "^11.8.1",
+    "@headlessui/react": "^1.6.4",
+    "@heroicons/react": "^1.0.6",
     "@mui/material": "^5.5.0",
     "@polygon.io/client-js": "^5.0.1",
     "@testing-library/jest-dom": "^5.16.2",
@@ -16,6 +18,8 @@
     "cross-env": "^7.0.3",
     "dotenv": "^16.0.0",
     "finnhub": "^1.2.10",
+    "headlessui": "^0.0.0",
+    "heroicons": "^1.0.6",
     "plotly.js": "^2.11.1",
     "react": "^17.0.2",
     "react-bootstrap": "^2.1.2",
@@ -53,6 +57,7 @@
     ]
   },
   "devDependencies": {
-    "prettier": "2.6.2"
+    "prettier": "2.6.2",
+    "tailwindcss": "^3.0.24"
   }
 }
